Separated gmailUtils file due to grpc package's multi thread bug

+++ src/data/schema/importHistory.ts
@@ -8,7 +8,7 @@ export const types = `
     contentType: String
     errorMsgs: [String]
     status: String
-    percentage: Int
+    percentage: Float
 
     date: Date
     user: User
